feat: [CDS-71173]: Added the setting control deleting execution (#52424)
* feat: [CDS-71173]: Added the setting control deleting pipeline executions

* feat: [CDS-71173]: Added the setting control deleting pipeline executions

* feat: [CDS-71173]: Added the setting control deleting pipeline executions

* feat: [CDS-71173]: review comments fix

* feat: [CDS-71173]: review comments fix

* feat: [CDS-71173]: review comments fix

* feat: [CDS-71173]: review comments fix

* feat: [CDS-71173]: code format issue

* feat: [CDS-71173]: code format issue

* feat: [CDS-71173]: code format issue

* feat: [CDS-71173]: UT fix

* feat: [CDS-71173]: UT fix

* feat: [CDS-71173]: UT fix

+++ pipeline-service/modules/orchestration-visualization/src/main/java/io/harness/event/PipelineExecutionSummaryDeleteObserver.java
@@ -25,10 +25,13 @@ public class PipelineExecutionSummaryDeleteObserver implements PlanExecutionDele
   @Inject PmsExecutionSummaryService pmsExecutionSummaryService;
 
   @Override
-  public void onPlanExecutionsDelete(List<PlanExecution> planExecutionList) {
-    Set<String> planExecutionIds = planExecutionList.stream().map(PlanExecution::getUuid).collect(Collectors.toSet());
+  public void onPlanExecutionsDelete(
+      List<PlanExecution> planExecutionList, boolean retainPipelineExecutionDetailsAfterDelete) {
+    if (!retainPipelineExecutionDetailsAfterDelete) {
+      Set<String> planExecutionIds = planExecutionList.stream().map(PlanExecution::getUuid).collect(Collectors.toSet());
 
-    // Delete all summary for given planExecutionIds
-    pmsExecutionSummaryService.deleteAllSummaryForGivenPlanExecutionIds(planExecutionIds);
+      // Delete all summary for given planExecutionIds
+      pmsExecutionSummaryService.deleteAllSummaryForGivenPlanExecutionIds(planExecutionIds);
+    }
   }
 }
